Change file name and subset samples to 100
In order to reduce processing time, I subsampled to 100 samples rather than the original 3942. The HGDP_1KG_PCALoadingsGenerate.py file name was also changed to follow the snake-case for filenames. Max-age was changed to '15h' to allow for more computational time in case the PCA function has not finished, and the gnomAD package was removed.