Change filtering order and moving liftover step to a new script
I reordered the filtering loci step so that it happens before joining datasets, as joins are computationally expensive so this should lower the processing time. I also created a new script (available in gnomad_90k_variants_liftover), which performs a liftover of the gnomad v2.1 loadings (from hg37 to hg38). As suggested by Vlad, I also rekeyed mt_tob_wgs, so that it has both the locus and alleles as its keys (and that way, has the same keys as mt_hgdp_1kg).