Filter tables 1 and 2 using a semi_join with loadings
I filtered tables 1 and 2 using a semi_join with loadings, then joined datasets by merging columns. As done previously, columns must be the same between datasets, so only 'GT' was selected, as this is used for the PCA step. In addition, I changed the outputdir option from gs://cpg-ancestry-temporary to gs://cpg-ancestry-analysis.