PCA and variant frequency analsis for NFE samples (#22)
* PCA and variant frequency analsis for NFE samples

This script performs a PCA on NFE samples only from the combined HGDP/1kG + TOB-WGS dataset. It also filters for variants which are in the gnomad loadings, but not in the TOB-WGS data.

* Change hgdp_1kg_path path extension from '.csv' to '.mt'.

* Change to temporary bucket

I can't write to the test bucket, only to the temporary bucket.

* Add in if __name__ == "__main__" statement

* Add in annotation to loadings rather than filtering HGDP/1kG + TOB-WGS matrix table

I added in gnomad AF, gnomad popmax AF, and whether variants are present in the TOB-WGS data into the loadings table.

* Compute and add TOB-WGS AF

I used the hl.agg.call_stats function to calculate the allele frequency on the TOB-WGS data, then added this into the loadings table. I also deleted the checkpointing of the filtered (nfe only) mt table, since it's easy to generate and not necessary to save.

* Update scripts/hail_batch/hgdp1kg_tobwgs_nfe/hgdp_1kg_tob_wgs_nfe.py

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>

* Update scripts/hail_batch/hgdp1kg_tobwgs_nfe/hgdp_1kg_tob_wgs_nfe.py

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>

* Add in gnomad_freq_meta information, remove TOB-variant column, and rename loadings mt

* Update scripts/hail_batch/hgdp1kg_tobwgs_nfe/hgdp_1kg_tob_wgs_nfe.py

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>