Add densify function before calculating AF for TOB-WGS data (#38)
* Add densify function before calculating AF for TOB-WGS data

I tested the densify function on a small subset of the TOB-WGS data, prior to calling the `lgt_to_gt` function. The AF changes after calling this function, and therefore I'm curious to see whether this has anything to do with the strange AF distribution we were seeing for the TOB-WGS data. This script is intended to get the AF after densifying the data, so that I can see how it affects the AF for the entire dataset.

* Update scripts/hail_batch/hgdp1kg_tobwgs_densify/hgdp_1kg_tob_wgs_densify.py

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>