Choose new variants for TOB-WGS + HGDP/1KG dataset
Since the ~90k variants that were previously chosen by gnomAD were not consistently capturing genotypes for the entire cohort, Loic and I went through the gnomAD v3 marker selection process, but applied to the combined TOB-WGS + HGDP/1kG data. By doing so, this should remove some of the batch effect we're seeing. The aim is to first filter out variants with an AF of 0.1%, but then see how the list changes as we increase this threshold.