Densify TOB-WGS data before joining HGDP/1kG and TOB-WGS datasets
I densified the TOB-WGS dataset before filtering and converting LGT to GT entries. We've found that densifying the sparse TOB-WGS mt is essential, as it affects downstream results. In order to compare the densified and non-densified matrix tables, I renamed the output files in this script by appending `_densified` to the end.