Choose new variants for TOB-WGS + HGDP/1KG dataset (#40)
* Choose new variants for TOB-WGS + HGDP/1KG dataset

Since the ~90k variants that were previously chosen by gnomAD were not consistently capturing genotypes for the entire cohort, Loic and I went through the gnomAD v3 marker selection process, but applied to the combined TOB-WGS + HGDP/1kG data. By doing so, this should remove some of the batch effect we're seeing. The aim is to first filter out variants with an AF of 0.1%, but then see how the list changes as we increase this threshold.

* Save final mt and densify

I forgot to save the final, filtered matrix variant table. I aso noticed that I did not densify the TOB-WGS matrix, which is now included as one of the initial steps. Finally, I removed the lines where we write out the joined HGDP/1kG + TOB-WGS matrix table.