Increase number of partitions and check number of rows from filtered mt
I increased the number of rows from 100 to 1000, since the previous script failed due to memory issues. Oddly, since the matrix table should be less than 90k rows, 100 partitions does seem sufficient, and therefore I added in a `count_rows()` statement to do a quick sanity check on how big the matrix table actually is. I also cached the tob_wgs data before printing the number of rows, so that I don't recompute the entire list of operations for printing and for saving the mt.