Change AF to 1% and subsample to 1M variants before pruning (#52)
* Change AF to 1% and subsample to 1M variants before pruning

After speaking with Loic this morning and looking at the histogram plot output from `hgdp_1kg_tob_wgs_variant_selection_exploration.py`, we decided that increasing the variant allele frequency to 0.01, rather than 0.001, would reduce many of the variants while still keeping rare variants in the analysis. We also found that ld_prune removes around 90% of the variants (at least on the subset of 1M variants that we tested on). Therefore, to retain around 100K, I've subsampled the remaining variants after filtering to 1M before LD pruning.

* Add test TOB-WGS dataset

Add test TOB-WGS dataset rather than the TOB-WGS dataset on main, as access will be denied due to read resctrictions on the test access level.

* Change from `test` to `main` bucket.

I tested the variant selection script on a subset of 1M variants, which finished successfully. I removed the subsetted data so that the analysis can be done on the entire dataset, as well as from the TOB-WGS data on `main`.

* Update scripts/hail_batch/variant_selection/hgdp_1kg_tob_wgs_variant_selection.py

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>

* Add constant for 1M rows

Co-authored-by: Leonhard Gruenschloss <leonhard.gruenschloss@populationgenomics.org.au>