Increase hdfs space for densified tob-wgs script (#55)
I changed the `main.py` script for the wrong folder (variant_selection rather than hgdp1kg_tobwgs_densified_pca). As stated in the PR cf2fd4c, my script failed with the error "File could only be replicated to 0 nodes instead of minReplication (=1)", and the GCP metrics explorer confirmed this was due to running out of HDFS space. I now increased this to 200.