Update README and job name
I updated the README to be more specific when referring to 'new variants'. I also updated the job name in the `hail_dataproc_job` function (the last one was a holdover) and removed the `key_rows_by` function from the tob_wgs matrix table, since the rows are already keyed by locus and allele.