Perform PCA on NFE subset, using new variants (#110)
* Test pop PCA for new variant selection

* Remove analysis_runner module

I was getting the error: 'ModuleNotFoundError: No module named 'analysis_runner' after running my last script, so I want to remove the analysis_runner to see if this is the only thing that's breaking.

* Include `install_common.sh` in `hail_dataproc_job` function

Not having this script is what was causing my previous error. I've gone ahead and reverted the script back to what it was previously, with the addition of adding this file in `main.py`.