Add in sampleid keys and fix allele annotation
Since there is a mismatch in keys within the genotype file and the other input files (i.e., the genotype files hace CPG IDs while the other files have OneK1K IDs), I got a TSV file of keys in order to translate between these. There are also differences in sampleid names between the keys that Seyhan gave me and the names within the files (the ones in the file have 6 digits, whereas the ones in the files have 3), so I've modifed the sampleids within the genotype files (with 6 digits) to mirror the others. Finally, I fixed an issue with adding in the alternative allele when there was only one allele.