Make sure expression data and covaraiate data are of equal length
I changed quite a bit in this last commit, but mostly it's around ensuring that the expression and covariate dfs are of equal length. I noticed that there were missing samples in both the covariate and expression dataframe, so I also added in a new metadata sheet which is more comprehensive so that I could get down to the bottom of why missing samples were arising.